+++ History.md
@@ -1,3 +1,8 @@
+unreleased
+==========
+
+  * Add generate option to control saving empty sessions
+
 1.3.1 / 2014-06-14
 ==================